% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_neuralforecast_env.R
\name{create_neuralforecast_env}
\alias{create_neuralforecast_env}
\title{Create Virtual Environment Wrapper}
\usage{
create_neuralforecast_env(envname = "nixtla-dev", ...)
}
\arguments{
\item{envname}{virtual environment to create}

\item{...}{additional passes for \code{create_virtualenv}}
}
\value{
creation of virtual environment
}
\description{
Create Virtual Environment Wrapper
}
