% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_neuralforecast.R
\name{install_neuralforecast}
\alias{install_neuralforecast}
\title{Install Development Version of neuralforecast From Github dev version contains conformal methods for predicting}
\usage{
install_neuralforecast(envname = "nixtla-dev", method = "auto", ...)
}
\arguments{
\item{envname}{virtual environment name}

\item{method}{method defaults to "auto"}

\item{...}{additional passes for \code{py_install}}
}
\value{
worth the wait! installs neuralforecast to "nixtla-dev" by default
}
\description{
Install Development Version of neuralforecast From Github dev version contains conformal methods for predicting
}
