% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_neuralforecast_env.R
\name{use_neuralforecast_env}
\alias{use_neuralforecast_env}
\title{Use Virtual Environment Wrapper}
\usage{
use_neuralforecast_env(envname = "nixtla-dev", ...)
}
\arguments{
\item{envname}{virtual environment to use}

\item{...}{additional passes for \code{use_virtualenv}}
}
\value{
sets environment to virtualenv
}
\description{
Use Virtual Environment Wrapper
}
